Break up the generator
