+++ src/blogmore/generator/constants.py
@@ -1,0 +1,24 @@
+"""Constants for the site generator."""
+
+CSS_FILENAME = "style.css"
+SEARCH_CSS_FILENAME = "search.css"
+STATS_CSS_FILENAME = "stats.css"
+ARCHIVE_CSS_FILENAME = "archive.css"
+CALENDAR_CSS_FILENAME = "calendar.css"
+GRAPH_CSS_FILENAME = "graph.css"
+TAG_CLOUD_CSS_FILENAME = "tag-cloud.css"
+GRAPH_JS_FILENAME = "graph.js"
+CODE_CSS_FILENAME = "code.css"
+THEME_JS_FILENAME = "theme.js"
+SEARCH_JS_FILENAME = "search.js"
+CODEBLOCKS_JS_FILENAME = "codeblocks.js"
+
+# All page-specific CSS files (source filenames only).
+PAGE_SPECIFIC_CSS: list[str] = [
+    SEARCH_CSS_FILENAME,
+    STATS_CSS_FILENAME,
+    ARCHIVE_CSS_FILENAME,
+    TAG_CLOUD_CSS_FILENAME,
+    CALENDAR_CSS_FILENAME,
+    GRAPH_CSS_FILENAME,
+]
